style: Rename utils to common
